Cine para cada pelicula con diferetes sesiones
